Изменил тип данных amount в products-dishes; готово заполнение таблицы products-dishes для первых 10 блюд.

+++ database/scripts/0_DB_creation.sql
@@ -86,7 +86,7 @@ CREATE TABLE "products" (
 CREATE TABLE "products_dishes" (
   "product_id" integer,
   "dish_id" integer,
-  "amount" integer,
+  "amount" numeric,
   PRIMARY KEY ("product_id", "dish_id")
 );
